# docs/swagger.yaml
definitions:
...
  dtos.BehaviourStreamsResponse:
    properties:
      countAlertLevel:
        type: integer
      countEvacuationLevel:
        type: integer
      countLowWaterLevel:
        type: integer
      totalValuesCount:
        type: integer
    type: object
  dtos.CalibratedStreamsData:
    properties:
      qualifier:
        type: string
      time:
        type: string
      value:
        type: number
    type: object
  dtos.CalibratedStreamsDataResponse:
    properties:
      mainStreams:
        items:
          $ref: '#/definitions/dtos.CalibratedStreamsData'
        type: array
      p05Streams:
        items:
          $ref: '#/definitions/dtos.CalibratedStreamsData'
        type: array
      p25Streams:
        items:
          $ref: '#/definitions/dtos.CalibratedStreamsData'
        type: array
      p75Streams:
        items:
          $ref: '#/definitions/dtos.CalibratedStreamsData'
        type: array
      p95Streams:
        items:
          $ref: '#/definitions/dtos.CalibratedStreamsData'
        type: array
    type: object
...
  dtos.ErrorResponse:
    properties:
      message:
        type: string
      timestamp:
        type: string
    type: object
...
  dtos.MetricCard:
    properties:
      name:
        type: string
      value:
        type: number
    type: object
...
  dtos.Pageable:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      pages:
        type: integer
      totalElements:
        type: integer
    type: object
...
  dtos.StreamCard:
    properties:
      checkErrors:
        type: boolean
      configuredStreamId:
        type: integer
      procedureId:
        type: integer
      procedureName:
        type: string
      stationId:
        type: integer
      stationName:
        type: string
      streamId:
        type: integer
      totalErrors:
        type: integer
      variableId:
        type: integer
      variableName:
        type: string
    type: object
  dtos.StreamCardsResponse:
    properties:
      content:
        items:
          $ref: '#/definitions/dtos.StreamCard'
        type: array
      pageable:
        $ref: '#/definitions/dtos.Pageable'
    type: object
  dtos.StreamData:
    properties:
      alertLevel:
        type: number
      evacuationLevel:
        type: number
      lowWaterLevel:
        type: number
      metrics:
        items:
          $ref: '#/definitions/dtos.MetricCard'
        type: array
      owner:
        type: string
      procId:
        type: integer
      procedure:
        type: string
      station:
        type: string
      streamType:
        type: integer
      unit:
        type: string
      unitId:
        type: integer
      updateFrequency:
        type: number
      varId:
        type: integer
      varName:
        type: string
    type: object
  dtos.StreamsData:
    properties:
      time:
        type: string
      value:
        type: number
    type: object
  dtos.StreamsDataResponse:
    properties:
      streams:
        items:
          $ref: '#/definitions/dtos.StreamsData'
        type: array
    type: object
...
  dtos.StreamsPerStation:
    properties:
      errorCount:
        type: integer
      stationId:
        type: string
      stationName:
        type: string
      streamsCount:
        type: integer
    type: object
  dtos.StreamsPerStationResponse:
    properties:
      stations:
        items:
          $ref: '#/definitions/dtos.StreamsPerStation'
        type: array
    type: object
...
info:
  contact: {}
  description: This API manages the inputs of the forecast model
  title: Inputs API
  version: "1.0"
paths:
...
  /series:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: mañana'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: Filtro por ID de la serie
        in: query
        name: streamId
        type: integer
      - description: Filtro por ID de la estacion
        in: query
        name: stationId
        type: integer
      - description: Filtro por ID de procedimiento
        in: query
        name: procId
        type: integer
      - description: Filtro por ID de variable
        in: query
        name: varId
        type: integer
      - description: Numero de pagina, por defecto 1
        in: query
        name: page
        type: integer
      - description: Cantidad de series por pagina, por defecto 15
        in: query
        name: pageSize
        type: integer
      - description: ID de la configuracion
        in: query
        name: configurationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamCardsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener las series configuradas de una configuracion
      tags:
      - Series
  /series/{serie_id}:
    get:
      parameters:
      - description: Id de la configuracion de la serie
        format: string
        in: query
        name: configuredStreamId
        required: true
        type: string
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: mañana'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: Id de la serie
        in: path
        name: serie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener los datos de una serie dado un id y su configuracion
      tags:
      - Series
  /series/comportamiento:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: mañana'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: ID de la configuracion
        in: query
        name: configurationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.BehaviourStreamsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener las metricas de comportamiento
      tags:
      - Series
  /series/curadas/{serie_id}:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: 5 dias despues'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: Id de la serie
        in: path
        name: serie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamsDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener los valores de una serie curada por id
      tags:
      - Series
  /series/estaciones:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: 5 dias despues'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: ID de la configuracion
        in: query
        name: configurationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamsPerStationResponse'
      summary: Endpoint para obtener el resumen de las series agrupado por estacion
      tags:
      - Series
  /series/observadas/{serie_id}:
    get:
      parameters:
      - description: 'Fecha de comienzo del periodo - valor por defecto: 7 dias atras'
        format: "2006-01-02"
        in: query
        name: timeStart
        type: string
      - description: 'Fecha del final del periodo - valor por defecto: mañana'
        format: "2006-01-02"
        in: query
        name: timeEnd
        type: string
      - description: Id de la serie
        in: path
        name: serie_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.StreamsDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener los valores de una serie observada por id
      tags:
      - Series
  /series/pronosticadas/{calibrado_id}:
    get:
      parameters:
      - description: Id del calibrado
        in: path
        name: calibrado_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CalibratedStreamsDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      summary: Endpoint para obtener los valores de una serie pronosticadas por id
      tags:
      - Series
...
swagger: "2.0"